fix(slicing): F_{K-1, Satterthwaite nu} reference for the joint period test
After the pairwise path moved to F_{1, Welch nu}, the joint omnibus in
the same file was the less-calibrated of the two: it read the diagonal
Wald statistic off an asymptotic chi-squared(K-1), measured at 5.7-6.7%
for a nominal 5% across K = 3-5 and T = 20-150 under a true null.

The two-sample Welch df is the K = 2 case of the Satterthwaite
approximation, and the omnibus is a diagonal-variance Wald form, so the
same approximation extends directly: nu = (sum v)^2 / sum(v^2 / (n-1))
on the variance shares (the scale-free form from #813's review), passed
to _wald_p_linear as df_denom. Measured 5.4-5.8% on the same nulls;
F_{K-1, min(n)-1} was considered and rejected as conservative
(4.2-4.6%). _welch_df is generalised to _satterthwaite_df and the
pairwise path calls the same function with two slices, so one helper
owns the df on both paths. reference_dist reads "f" and df_denom carries
nu on the analytic omnibus; the bootstrap omnibus is unchanged.

_wald_p_linear's df_denom is typed float | None — the Satterthwaite nu
is fractional, and scipy's f.sf accepts it.

Breaking: analytic slice_period_joint_test p-values move toward
calibrated; reference_dist / df_denom values change on that path.

+++ factrix/_stats/wald.py
@@ -43,7 +43,7 @@ def _wald_p_linear(
     R: np.ndarray,
     q: np.ndarray | float = 0.0,
     *,
-    df_denom: int | None = None,
+    df_denom: float | None = None,
 ) -> tuple[float, float]:
     """Wald test of the linear restriction ``Rβ = q``.
